+++ ProfilesAPI/Application/Services/PatientsFolder/Queries/ViewPatientProfile/ViewPatientProfileQueryHandler.cs
@@ -14,9 +14,9 @@ public class ViewPatientProfileQueryHandler(IPatientsRepo _patientsRepo) : IRequ
         var patient = await _patientsRepo.GetPatientById(
             request.IdPatient, cancellationToken);
         
-        if (patient == null) return new CustomResult(false, Messages.PatientNotFound, (int)HttpStatusCode.NotFound);
+        if (patient == null) return new CustomResult(false, Messages.PatientNotFound, HttpStatusCode.NotFound);
 
         var patientReadDto = PatientReadDto.MapFromPatient(patient);
-        return new CustomResult(true, Messages.Success, (int)HttpStatusCode.OK, patientReadDto);
+        return new CustomResult(true, Messages.Success, HttpStatusCode.OK, patientReadDto);
     }
 }
